Bug fix for issue #4
closes gh-4

No descriptor pool was being created, resulting in a error when trying to create a new descriptor set.
Since the shaders need the descriptor sets for the lights, now one is created for each object.
This seems like overkill though (new issue created #23).
Now there is a validation error because we are binding a descriptor that (new issue created #24).